fix(review): consolidate SKILL_SEARCH_DIRS and _resolve_skill_md into _skill_helpers.py
Eliminates duplication between rules_stamp_ownership.py and
rules_skill_content.py by extracting the shared test-injection
seam and skill resolution helper to recipe/_skill_helpers.py.
